+++ internal/data/users.go
@@ -23,6 +23,7 @@ type Users interface {
 	GetByGithubId(githubId int64) (*User, error)
 
 	FilterById(id *int64) Users
+	FilterByLowerTime(time time.Time) Users
 	SearchBy(search string) Users
 
 	Page(pageParams pgdb.OffsetPageParams) Users
@@ -34,6 +35,7 @@ type User struct {
 	GithubId  int64     `json:"id" db:"github_id" structs:"github_id"`
 	AvatarUrl string    `json:"avatar_url" db:"avatar_url" structs:"avatar_url"`
 	CreatedAt time.Time `json:"created_at" db:"created_at" structs:"-"`
+	UpdatedAt time.Time `json:"updated_at" db:"updated_at" structs:"-"`
 }
 
 type UnverifiedUser struct {