+++ internal/github/add_update_user.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"time"
 
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 	"gitlab.com/distributed_lab/acs/github-module/internal/data"
@@ -49,53 +50,6 @@ func (g *github) AddUserFromApi(link, username, permission string) (*data.Permis
 	}
 }
 
-func (g *github) UpdateUserFromApi(link, username, permission string) (*data.Permission, error) {
-	findType, _, err := g.FindType(link)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get type")
-	}
-
-	if err = validation.Validate(findType, validation.In(data.Organization, data.Repository)); err != nil {
-		return nil, errors.Wrap(err, "something wrong with link type")
-	}
-
-	switch findType {
-	case data.Repository:
-		isCollaborator, _, err := g.CheckRepoCollaborator(link, username)
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to check if user in repo from api")
-		}
-
-		if !isCollaborator {
-			return nil, errors.Errorf("such user is not in repository")
-		}
-
-		owned, err := g.FindRepoOwner(link)
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to check if repository owner")
-		}
-
-		if owned == data.UserOwned {
-			permission = "write"
-		}
-
-		return g.AddOrUpdateUserInRepoFromApi(link, username, permission)
-	case data.Organization:
-		isCollaborator, _, err := g.CheckOrgCollaborator(link, username)
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to check if user in org from api")
-		}
-
-		if !isCollaborator {
-			return nil, errors.Errorf("`%s` is not in organisation `%s`", username, link)
-		}
-
-		return g.AddOrUpdateUserInOrgFromApi(link, username, permission)
-	default:
-		return nil, errors.Errorf("unexpected type `%s`", findType)
-	}
-}
-
 func (g *github) AddOrUpdateUserInRepoFromApi(link, username, permission string) (*data.Permission, error) {
 	jsonBody, _ := json.Marshal(struct {
 		Permission string `json:"permission"`
@@ -119,6 +73,16 @@ func (g *github) AddOrUpdateUserInRepoFromApi(link, username, permission string)
 
 	if res == nil {
 		return nil, errors.New("failed to process request: response is nil")
+	}
+
+	if res.StatusCode == http.StatusForbidden {
+		timeoutDuration, err := g.getDuration(res)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to get time duration from response")
+		}
+		g.log.Warnf("we need to wait `%d`", timeoutDuration)
+		time.Sleep(timeoutDuration)
+		return g.AddOrUpdateUserInRepoFromApi(link, username, permission)
 	}
 
 	if res.StatusCode < 200 || res.StatusCode >= 300 {
@@ -185,6 +149,16 @@ func (g *github) AddOrUpdateUserInOrgFromApi(link, username, permission string) 
 
 	if res == nil {
 		return nil, errors.New("failed to process request: response is nil")
+	}
+
+	if res.StatusCode == http.StatusForbidden {
+		timeoutDuration, err := g.getDuration(res)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to get time duration from response")
+		}
+		g.log.Warnf("we need to wait `%d`", timeoutDuration)
+		time.Sleep(timeoutDuration)
+		return g.AddOrUpdateUserInOrgFromApi(link, username, permission)
 	}
 
 	if res.StatusCode < 200 || res.StatusCode >= 300 {