+++ internal/github/get_users.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
+	"time"
 
 	"gitlab.com/distributed_lab/acs/github-module/internal/data"
 	"gitlab.com/distributed_lab/logan/v3/errors"
@@ -28,13 +29,24 @@ func (g *github) GetUsersFromApi(link, typeTo string) ([]data.Permission, error)
 
 	for page := 1; ; page++ {
 		q := req.URL.Query()
-		q.Add("per_page", "100")
-		q.Add("page", strconv.Itoa(page))
+		q.Set("per_page", "100")
+		q.Set("page", strconv.Itoa(page))
 		req.URL.RawQuery = q.Encode()
 
 		res, err := http.DefaultClient.Do(req)
 		if err != nil {
 			return nil, errors.Wrap(err, " error making http request")
+		}
+
+		if res.StatusCode == http.StatusForbidden {
+			timeoutDuration, err := g.getDuration(res)
+			if err != nil {
+				return nil, errors.Wrap(err, "failed to get time duration from response")
+			}
+			g.log.Warnf("we need to wait `%d`", timeoutDuration)
+			time.Sleep(timeoutDuration)
+			page--
+			continue
 		}
 
 		if res.StatusCode < 200 || res.StatusCode >= 300 {