+++ internal/processor/add_user.go
@@ -78,7 +78,7 @@ func (p *processor) handleAddUserAction(msg data.ModulePayload) error {
 		return errors.Wrap(err, "failed to make add user transaction")
 	}
 
-	err = p.sendDeleteUser(msg.RequestId, dbUser)
+	err = p.SendDeleteUser(msg.RequestId, dbUser)
 	if err != nil {
 		p.log.WithError(err).Errorf("failed to publish users for message action with id `%s`", msg.RequestId)
 		return errors.Wrap(err, "failed to publish users")