+++ CHANGELOG.md
@@ -6,7 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.0.0] - 2023-03-05
+## [1.0.3] - 2023-03-27
+
+### Added
+
+- Worker delete users and permissions that wasn't updated for some time
+- `updated_at` column in users table
+
+## [1.0.2] - 2023-03-22
+
+### Added
+
+- Request to send roles by user status
+
+## [1.0.1] - 2023-03-17
+
+### Added
+
+- Handling `Too Many requests` error from `API`
+
+## [1.0.0] - 2023-03-15
 
 ### Added
 
@@ -15,5 +34,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Database.
 - API handlers.
 
-[1.1.0]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.0...v1.1.0
 [1.0.0]: https://gitlab.com/distributed_lab/acs/github-module/-/tree/feature/review_fixes
+[1.0.1]: https://gitlab.com/distributed_lab/acs/gitlab-module/-/compare/feature/review_fixes...feature/many_requests_handle
+[1.0.2]: https://gitlab.com/distributed_lab/acs/gitlab-module/-/compare/feature/review_fixes...feature/many_requests_handle